Use psql float instead of bigint to avoid overflow

+++ prisma/sql/getProvidersWeeklyAcc.sql
@@ -10,8 +10,8 @@ with "provider_retrievability_weekly" as (select date_trunc('week', "date")     
 select "week"                                as "week",
        "provider"                            as "provider",
        count(*)::int                         as "num_of_clients",
-       max("total_deal_size")::bigint        as "biggest_client_total_deal_size",
-       sum("total_deal_size")::bigint        as "total_deal_size",
+       max("total_deal_size")::float         as "biggest_client_total_deal_size",
+       sum("total_deal_size")::float         as "total_deal_size",
        max(coalesce("success_rate", 0))      as "avg_retrievability_success_rate",
        max(coalesce("success_rate_http", 0)) as "avg_retrievability_success_rate_http"
 from "client_provider_distribution_weekly_acc" as "cpdwa"
